Feat: apply logic + style site w/ Dart
Functionally similar to the TypeScript version except that random rgb
colors are missing when hovering over. Could not apply
'style.background' on target with dart:web package. No idea if it isn't
available or a different approach is needed. According to older SO posts
this is available in legacy Dart:html package